Added a generic HTTP Request to avoid duplicate codes & Put all weather related API Call in the same Data Service

+++ TreeOP/Models/AirQuality.swift
@@ -13,7 +13,7 @@ struct AirQualityResponse: Decodable {
 }
 
 struct AQData: Decodable {
-    let aqIndex: Double
+    let aqIndex: Int
     
     enum CodingKeys: String, CodingKey {
         case aqIndex = "aqi"
